// src/app/modules/master/models/banner-detail.ts
export class BannerData {
        title: string;
        banner: string;
        order: number;
        start_date: string;
        end_date: string;
        clickable_flg: boolean;
        clickable_is_detail: boolean;
        clickable_is_internal: boolean;
        clickable_redirect: string;
        description: string;
        foot_text_flg: string;
        foot_text_content: string;
        foot_text_redirect: string;
        foot_image_flg: string;
        foot_image_content: string;
        foot_button_redirect: string;
        foot_button_flg: string;
        foot_button_content: string;
        foot_image_redirect: string;
}

export class UpdateBannerData {
        id: string;
        title: string;
        banner: string;
        order: number;
        start_date: string;
        end_date: string;
        clickable_flg: boolean;
        clickable_is_detail: boolean;
        clickable_is_internal: boolean;
        clickable_redirect: string;
        description: string;
        foot_text_flg: string;
        foot_text_content: string;
        foot_text_redirect: string;
        foot_image_flg: string;
        foot_image_content: string;
        foot_button_redirect: string;
        foot_button_flg: string;
        foot_button_content: string;
        foot_image_redirect: string;
}