// src/app/shared/models/lov.ts
export class LovData {
        id: string;
        name: string;
        value: string;
        description: string;
        par_id: string;
        lov_type_id: string;
        strvalue_1: string;
        strvalue_2: string;
        strvalue_3: string;
        numvalue_1: string;
        numvalue_2: string;
        numvalue_3: string;
        created_dt: string;
        created_by: string;
        modified_dt: string;
        modified_by: string;
        is_deleted: boolean;
}